Fixes to loop variable and various problems in simd routine
Fixed a bug with the loop variable generation at higher opt levels. Added some placeholders for distribute and for loop generation, but some extra work needs to be figured out from the code generation side. Fixed a lot of unneeded aspects in the code generation for simd loops. Added support for private and firstprivate for simd loops.