Cleanup and testing
Synchronization done in the commmon plugin interface. Put a check to make sure the asyncinfo is valid, since there is some data exchange that happens at the beginning of the kernel without valid async info. This does add a check before each offloaded operation for the environemnt variable, and only if that var is set does it do the synchronization.